Target update

+++ components/ld2450/ld2450.h
@@ -25,7 +25,7 @@ struct Target
 {
     int16_t x;
     int16_t y;
-    int16_t speed;
+    int speed;
     uint16_t resolution;
 };
